Code review returns: Thomas 2
